Reworking extensions, again
