+++ module/auth-rest/extension/context.rest.interceptor.ts
@@ -1,0 +1,5 @@
+import { AuthInterceptor } from '..';
+
+const { ContextInterceptor } = require('@travetto/context/extension/rest.interceptor');
+const arr = (AuthInterceptor.prototype.after = AuthInterceptor.prototype.after || []) as any[];
+arr.push(ContextInterceptor);
