+++ module/auth-rest/package.json
@@ -26,7 +26,8 @@
   "main": "index.ts",
   "files": [
     "index.ts",
-    "src"
+    "src",
+    "extension"
   ],
   "name": "@travetto/auth-rest",
   "repository": {
