+++ module/context/package.json
@@ -28,7 +28,8 @@
   "main": "index.ts",
   "files": [
     "index.ts",
-    "src"
+    "src",
+    "extension"
   ],
   "name": "@travetto/context",
   "repository": {
