+++ module/email/package.json
@@ -27,7 +27,8 @@
   "main": "index.ts",
   "files": [
     "index.ts",
-    "src"
+    "src",
+    "extension"
   ],
   "name": "@travetto/email",
   "repository": {
