+++ module/jwt/package.json
@@ -5,7 +5,8 @@
   "main": "index.ts",
   "files": [
     "index.ts",
-    "src"
+    "src",
+    "extension"
   ],
   "repository": {
     "type": "git",
