+++ module/model-mongo/test/simple.ts
@@ -4,7 +4,7 @@ import { Model, ModelService, BaseModel, ModelSource } from '@travetto/model';
 import { DependencyRegistry } from '@travetto/di';
 import { Suite, Test } from '@travetto/test';
 import { Schema, Min, Max } from '@travetto/schema';
-import { GenerateUtil } from '@travetto/schema/src/extension/faker.ext';
+import { GenerateUtil } from '@travetto/schema/extension/faker';
 
 import { MongoModelSource } from '../';
 import { BaseMongoTest } from './base';
