+++ module/rest-session/package.json
@@ -24,7 +24,8 @@
   "main": "index.ts",
   "files": [
     "index.ts",
-    "src"
+    "src",
+    "extension"
   ],
   "name": "@travetto/rest-session",
   "repository": {
