+++ module/rest/index.ts
@@ -15,10 +15,11 @@ export * from './src/registry/types';
 export * from './src/response/redirect';
 export * from './src/response/renderable';
 export * from './src/response/string';
-export * from './src/interceptor/group';
+export * from './src/interceptor/all';
 export * from './src/interceptor/cors';
 export * from './src/interceptor/get-cache';
-export * from './src/interceptor/types';
+export * from './src/interceptor/group';
+export * from './src/interceptor/interceptor';
 export * from './src/types';
 export * from './src/util/mime';
 export * from './src/util/rest-app';
