+++ module/rest/src/interceptor/group.ts
@@ -1,25 +1,46 @@
-import { Injectable, DependencyRegistry } from '@travetto/di';
 import { Class } from '@travetto/registry';
-import { AppInfo } from '@travetto/base';
+import { DependencyRegistry } from '@travetto/di';
+import { Util } from '@travetto/base';
 
-import { RestInterceptorGroup, RestInterceptor } from './types';
+import { RestInterceptor } from './interceptor';
 
-@Injectable({ target: RestInterceptorGroup })
-export class AllInterceptorGroup extends RestInterceptorGroup {
+export class RestInterceptorGroup {
+  public interceptors: Set<Class<RestInterceptor>>;
+  constructor(...interceptors: Class<RestInterceptor>[]) {
+    this.interceptors = new Set(interceptors);
+  }
 
-  async postConstruct() {
-    const items = DependencyRegistry.getCandidateTypes(RestInterceptor as Class);
+  async getActive() {
+    // Final all available interceptors
+    const interceptors = DependencyRegistry.getCandidateTypes(RestInterceptor as Class)
+      .filter(x => this.interceptors.has(x.class));
 
-    const out: Class<RestInterceptor>[] = [];
-    for (const item of items) {
-      const file = item.class.__filename;
-      let target = AppInfo.NAME;
-      if (file.includes('node_modules')) {
-        target = file.replace(/^.*(@travetto\/[^\/]+).*/, (a, key) => key);
-      }
-      out.push(item.class as Class<RestInterceptor>);
-    }
+    // Get instances for all of them
+    const instances = await Promise.all<RestInterceptor>(interceptors.map(op =>
+      DependencyRegistry.getInstance(op.target, op.qualifier)
+        .catch(err => {
+          if ((err.message || '').includes('Cannot find module')) {
+            console.error(`Unable to load operator ${op.class.name}#${op.qualifier.toString()}, module not found`);
+          } else {
+            throw err;
+          }
+        })
+    ));
 
-    this.interceptors = new Set(out);
+    // Sort according to before/after
+    const sorted = Util.computeOrdering(
+      instances
+        .map(x => ({
+          key: x.constructor,
+          before: x.before,
+          after: x.after,
+          target: x
+        }))
+    )
+      .map(x => x.target);
+
+    console.debug('Sorting interceptors', sorted.length, sorted.map(x => x.constructor.name));
+
+    return sorted;
   }
 }
